feat: Backend init refactored
